% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{Launches the studydashboard}
\usage{
launchApp()
}
\description{
This function launches the studydashboard in your console.
}
\seealso{
\code{\link{launchApp_background}}, \code{\link{reset_db}}
}
