% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_db.R
\name{reset_db}
\alias{reset_db}
\title{Delete Database Content}
\usage{
reset_db(fs = TRUE, studium = TRUE, pruefung = TRUE)
}
\arguments{
\item{fs}{Logical wich defines if the fs data should be deleted.}

\item{studium}{Logical wich defines if the studium data should be deleted.}

\item{pruefung}{Logical wich defines if the pruefung data should be deleted.}
}
\description{
This function deletes the database content.
}
\examples{
# Delete fs and pruefung data:
reset_db(studium = FALSE)
}
\seealso{
\code{\link{launchApp}}, \code{\link{launchApp_background}}
}
