% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studydashboard.R
\docType{package}
\name{studydashboard}
\alias{studydashboard}
\title{studydashboard: A Package Containing the Studydashboard}
\description{
This package provides functions to launch and maintain the studydashboard application. The main focus of this application lies on the visualisation of earned so called Credit Points in the European Credit Transfer System (in short ECTSCP) over the course of students’ studies. For this we offer a user interace which makes it user-friendly to produce such graphics.
}
\details{
The studydashboard package provides three important functions:

launchApp, launchApp_background and reset_db.
}
\section{studydashboard functions}{

\code{\link{launchApp}},
\code{\link{launchApp_background}},
\code{\link{reset_db}}
}

